refactor(#270): extract shared per-step renderer to eliminate emitter divergence
PR #274 review surfaced a recurring class of bugs caused by the
template emitter re-implementing logic that already lives correctly
in the per-endpoint Playwright emitter. Three review rounds found
five distinct divergences:

  1. multi-extract redeclaration of `const jsonN`
  2. dotted `elementField` lookup
  3. body-template placeholder regex (\${var} vs ${var})
  4. eventually-consistent ops not wrapped in awaitEventually()
  5. URL-param fallback shape

Patching each instance one-at-a-time has not converged — each round
finds the next divergence. This commit addresses the *class* by
extracting the shared primitives into materializer/src/playwright/
stepRenderer.ts:

  - BODY_PLACEHOLDER_RE (accepts both "\${var}" and "${var}")
  - buildUrlExpression / camelCase / escapeQuotes / toOptionalAccessor
  - renderInlineStepLines (URL + body + multipart + auth + EC wrap
    + status assertion + error logging)
  - renderEventualWait (#159 PR B wait block)
  - stepNeedsAwaitForOp (per-step EC-wrap predicate, parameterised on
    a Set<string> so both emitters can compute the set their own way)
  - reindent (helper for embedding shared blocks in deeper scopes)

Both emitters now call into these primitives. The per-endpoint
emitter's output is byte-identical (verified by diffing all 261
.spec.ts files before/after this refactor); only the template
emitter's output changes — picking up the fixes for divergences 3,
4, and 5.

Other changes:

* TemplateScenario gains `eventuallyConsistentOps: string[]`,
  populated by scenarioTemplateInstantiator from each step
  operation's OperationSpec.eventuallyConsistent flag. The template
  emitter consumes this set to drive stepNeedsAwaitForOp() — without
  it, the template emitter has no way to know which read steps need
  EC wrapping at emit time.

* path-analyser/src/index.ts: wipe the entire
  generated/<config>/scenarios/templates/ tree (via the new
  getTemplateScenariosRootDir accessor) rather than only the dirs
  for the templates currently producing results. Previously, a
  template removed from the ABox (or whose appliesTo changed to
  match no subject) would leave a stale per-template directory and
  its per-subject JSON files behind across runs. (#270 review.)

Full gate green: lint clean, all 7 typechecks pass, per-endpoint
output byte-identical, 538 passed | 7 skipped.
